+++ lib/promauth/config.go
@@ -2,15 +2,19 @@ package promauth
 
 import (
 	"bytes"
+	"context"
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/base64"
 	"fmt"
 	"io/ioutil"
+	"net/url"
 	"sync"
 
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/fasttime"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/logger"
+	"golang.org/x/oauth2"
+	"golang.org/x/oauth2/clientcredentials"
 )
 
 // TLSConfig represents TLS config.
@@ -46,6 +50,7 @@ type HTTPClientConfig struct {
 	BasicAuth       *BasicAuthConfig `yaml:"basic_auth,omitempty"`
 	BearerToken     string           `yaml:"bearer_token,omitempty"`
 	BearerTokenFile string           `yaml:"bearer_token_file,omitempty"`
+	OAuth2          *OAuth2Config    `yaml:"oauth2,omitempty"`
 	TLSConfig       *TLSConfig       `yaml:"tls_config,omitempty"`
 }
 
@@ -56,6 +61,97 @@ type ProxyClientConfig struct {
 	BearerToken     string           `yaml:"proxy_bearer_token,omitempty"`
 	BearerTokenFile string           `yaml:"proxy_bearer_token_file,omitempty"`
 	TLSConfig       *TLSConfig       `yaml:"proxy_tls_config,omitempty"`
+}
+
+// OAuth2Config represent OAuth2 configuration
+type OAuth2Config struct {
+	ClientID         string            `yaml:"client_id"`
+	ClientSecret     string            `yaml:"client_secret"`
+	ClientSecretFile string            `yaml:"client_secret_file"`
+	Scopes           []string          `yaml:"scopes"`
+	TokenURL         string            `yaml:"token_url"`
+	EndpointParams   map[string]string `yaml:"endpoint_params"`
+}
+
+// String returns string representation of o.
+func (o *OAuth2Config) String() string {
+	return fmt.Sprintf("clientID=%q, clientSecret=%q, clientSecretFile=%q, Scopes=%q, tokenURL=%q, endpointParams=%q",
+		o.ClientID, o.ClientSecret, o.ClientSecretFile, o.Scopes, o.TokenURL, o.EndpointParams)
+}
+
+func (o *OAuth2Config) validate() error {
+	if o.ClientID == "" {
+		return fmt.Errorf("client_id cannot be empty")
+	}
+	if o.ClientSecret == "" && o.ClientSecretFile == "" {
+		return fmt.Errorf("ClientSecret or ClientSecretFile must be set")
+	}
+	if o.ClientSecret != "" && o.ClientSecretFile != "" {
+		return fmt.Errorf("ClientSecret and ClientSecretFile cannot be set simultaneously")
+	}
+	if o.TokenURL == "" {
+		return fmt.Errorf("token_url cannot be empty")
+	}
+	return nil
+}
+
+type oauth2ConfigInternal struct {
+	mu               sync.Mutex
+	cfg              *clientcredentials.Config
+	clientSecretFile string
+	tokenSource      oauth2.TokenSource
+}
+
+func newOAuth2ConfigInternal(baseDir string, o *OAuth2Config) (*oauth2ConfigInternal, error) {
+	if err := o.validate(); err != nil {
+		return nil, err
+	}
+	oi := &oauth2ConfigInternal{
+		cfg: &clientcredentials.Config{
+			ClientID:       o.ClientID,
+			ClientSecret:   o.ClientSecret,
+			TokenURL:       o.TokenURL,
+			Scopes:         o.Scopes,
+			EndpointParams: urlValuesFromMap(o.EndpointParams),
+		},
+	}
+	if o.ClientSecretFile != "" {
+		oi.clientSecretFile = getFilepath(baseDir, o.ClientSecretFile)
+		secret, err := readPasswordFromFile(oi.clientSecretFile)
+		if err != nil {
+			return nil, fmt.Errorf("cannot read OAuth2 secret from %q: %w", oi.clientSecretFile, err)
+		}
+		oi.cfg.ClientSecret = secret
+	}
+	oi.tokenSource = oi.cfg.TokenSource(context.Background())
+	return oi, nil
+}
+
+func urlValuesFromMap(m map[string]string) url.Values {
+	result := make(url.Values, len(m))
+	for k, v := range m {
+		result[k] = []string{v}
+	}
+	return result
+}
+
+func (oi *oauth2ConfigInternal) getTokenSource() (oauth2.TokenSource, error) {
+	oi.mu.Lock()
+	defer oi.mu.Unlock()
+
+	if oi.clientSecretFile == "" {
+		return oi.tokenSource, nil
+	}
+	newSecret, err := readPasswordFromFile(oi.clientSecretFile)
+	if err != nil {
+		return nil, fmt.Errorf("cannot read OAuth2 secret from %q: %w", oi.clientSecretFile, err)
+	}
+	if newSecret == oi.cfg.ClientSecret {
+		return oi.tokenSource, nil
+	}
+	oi.cfg.ClientSecret = newSecret
+	oi.tokenSource = oi.cfg.TokenSource(context.Background())
+	return oi.tokenSource, nil
 }
 
 // Config is auth config.
@@ -131,16 +227,16 @@ func (ac *Config) NewTLSConfig() *tls.Config {
 
 // NewConfig creates auth config for the given hcc.
 func (hcc *HTTPClientConfig) NewConfig(baseDir string) (*Config, error) {
-	return NewConfig(baseDir, hcc.Authorization, hcc.BasicAuth, hcc.BearerToken, hcc.BearerTokenFile, hcc.TLSConfig)
+	return NewConfig(baseDir, hcc.Authorization, hcc.BasicAuth, hcc.BearerToken, hcc.BearerTokenFile, hcc.OAuth2, hcc.TLSConfig)
 }
 
 // NewConfig creates auth config for the given pcc.
 func (pcc *ProxyClientConfig) NewConfig(baseDir string) (*Config, error) {
-	return NewConfig(baseDir, pcc.Authorization, pcc.BasicAuth, pcc.BearerToken, pcc.BearerTokenFile, pcc.TLSConfig)
+	return NewConfig(baseDir, pcc.Authorization, pcc.BasicAuth, pcc.BearerToken, pcc.BearerTokenFile, nil, pcc.TLSConfig)
 }
 
 // NewConfig creates auth config from the given args.
-func NewConfig(baseDir string, az *Authorization, basicAuth *BasicAuthConfig, bearerToken, bearerTokenFile string, tlsConfig *TLSConfig) (*Config, error) {
+func NewConfig(baseDir string, az *Authorization, basicAuth *BasicAuthConfig, bearerToken, bearerTokenFile string, o *OAuth2Config, tlsConfig *TLSConfig) (*Config, error) {
 	var getAuthHeader func() string
 	authDigest := ""
 	if az != nil {
@@ -229,6 +325,29 @@ func NewConfig(baseDir string, az *Authorization, basicAuth *BasicAuthConfig, be
 			return "Bearer " + bearerToken
 		}
 		authDigest = fmt.Sprintf("bearer(token=%q)", bearerToken)
+	}
+	if o != nil {
+		if getAuthHeader != nil {
+			return nil, fmt.Errorf("cannot simultaneously use `authorization`, `basic_auth, `bearer_token` and `ouath2`")
+		}
+		oi, err := newOAuth2ConfigInternal(baseDir, o)
+		if err != nil {
+			return nil, err
+		}
+		getAuthHeader = func() string {
+			ts, err := oi.getTokenSource()
+			if err != nil {
+				logger.Errorf("cannot get OAuth2 tokenSource: %s", err)
+				return ""
+			}
+			t, err := ts.Token()
+			if err != nil {
+				logger.Errorf("cannot get OAuth2 token: %s", err)
+				return ""
+			}
+			return t.Type() + " " + t.AccessToken
+		}
+		authDigest = fmt.Sprintf("oauth2(%s)", o.String())
 	}
 	var tlsRootCA *x509.CertPool
 	var tlsCertificate *tls.Certificate